+++ pmd-core/src/main/java/net/sourceforge/pmd/internal/util/FileCollectionUtil.java
@@ -7,7 +7,6 @@ package net.sourceforge.pmd.internal.util;
 import java.io.IOException;
 import java.io.Reader;
 import java.net.URI;
-import java.nio.file.FileSystem;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.sql.SQLException;
@@ -42,11 +41,29 @@ public final class FileCollectionUtil {
             collector.setCharset(configuration.getSourceEncoding());
         }
 
-        collectFiles(collector, configuration.getInputPathList());
+
+        // This is to be removed when --short-names is removed.
+        // If the new --relativize-paths-with option is specified (!= null), it takes precedence.
+        boolean legacyShortNamesBehavior =
+                configuration.getRelativizeRoots().isEmpty() && configuration.isReportShortNames();
+
+        for (Path path : configuration.getInputPathList()) {
+            try {
+                if (legacyShortNamesBehavior) {
+                    collector.relativizeWith(path.toString());
+                }
+                addRoot(collector, path);
+            } catch (IOException e) {
+                collector.getReporter().errorEx("Error collecting " + path, e);
+            }
+        }
+        // use that, once --short-names is removed
+        //collectFiles(collector, configuration.getInputPathList());
 
         if (configuration.getUri() != null) {
             collectDB(collector, configuration.getUri());
         }
+
 
         if (configuration.getInputFile() != null) {
             collectFileList(collector, configuration.getInputFile());
@@ -104,15 +121,7 @@ public final class FileCollectionUtil {
             LOG.debug("Adding directory {}.", path);
             collector.addDirectory(path);
         } else if (pathStr.endsWith(".zip") || pathStr.endsWith(".jar")) {
-            LOG.debug("Adding zip file {}.", path);
-            @SuppressWarnings("PMD.CloseResource")
-            FileSystem fs = collector.addZipFile(path);
-            if (fs == null) {
-                return;
-            }
-            for (Path zipRoot : fs.getRootDirectories()) {
-                collector.addFileOrDirectory(zipRoot);
-            }
+            collector.addZipFileWithContent(path);
         } else if (Files.isRegularFile(path)) {
             LOG.debug("Adding regular file {}.", path);
             collector.addFile(path);